Move extension traits away from libsqlite3_sys
